telegram_bot/pyproject.toml
[project]
name = "telegram_bot"
version = "0.1.0"
dependencies = [
    "aiogram==3.23.0",
    "httpx==0.28.1",
    "python-dotenv==1.2.1",
    "redis==7.1.0",
    "timezonefinder==8.1.0",
    "numpy==2.3.0",
]

[project.optional-dependencies]
dev = [
]
